requirements.txt
ollama
fastapi
psycopg2
pygithub
